# Optimize mine-tokens execution costs