--- a/src/citycoin-client.ts
+++ b/src/citycoin-client.ts
@@ -121,12 +121,12 @@ export class CityCoinClient {
     )
   }
 
-  hasMinedInList(miner: Account, miners: MinersList): Result {
+  hasMined(miner: Account, blockHeight: number): Result {
     return this.callReadOnlyFn(
-      "has-mined-in-list",
+      "has-mined",
       [
         types.principal(miner.address),
-        miners.convert()
+        types.uint(blockHeight)
       ]
     );
   }