This PR contains changes that optimize cost of execution `mine-tokens` function.

I replaced very expensive `has-mined-in-list` function that uses `fold` to simulate loop and check if miner already mined in selected block or not, with much faster and least expensive `has-mined` which uses `map-get` to achieve the same result.

As a consequence of this modification I had to introduce new map `miners-block-commitment` which stores information if miner committed anything at specific block height or not.

Execution costs were acquired as follows:
```sh
$ npm run clarinet:console
> (contract-call? .citycoin register-miner)
> ::advance_chain_tip 200
> ::get_costs (contract-call? .citycoin mine-tokens u200)
```

**BEFORE**
```
+----------------------+-----------+------------+
|                      | Consumed  | Limit      |
+----------------------+-----------+------------+
| Runtime              | 316336000 | 5000000000 |
+----------------------+-----------+------------+
| Read count           | 17        | 7750       |
+----------------------+-----------+------------+
| Read length (bytes)  | 45591     | 100000000  |
+----------------------+-----------+------------+
| Write count          | 3         | 7750       |
+----------------------+-----------+------------+
| Write length (bytes) | 6875      | 15000000   |
+----------------------+-----------+------------+
```

**AFTER**
```
+----------------------+----------+------------+
|                      | Consumed | Limit      |
+----------------------+----------+------------+
| Runtime              | 61794000 | 5000000000 |
+----------------------+----------+------------+
| Read count           | 19       | 7750       |
+----------------------+----------+------------+
| Read length (bytes)  | 45365    | 100000000  |
+----------------------+----------+------------+
| Write count          | 4        | 7750       |
+----------------------+----------+------------+
| Write length (bytes) | 7122     | 15000000   |
+----------------------+----------+------------+
```

Overall this change reduced `mine-tokens` execution costs  ~5.11x.

@jcnelson if you'll have a time, I would love to hear your opinion about changes like this. Do you see any potential issues they might cause?